Fixed interaction with historical plan year filter.
What's happening here is that earlier in the process we remove nodes
from the doc that might be REALLY OLD plan years.  Then later we fail
to check if those have resulted in an empty event set.

Currently in DC we have some noise reduction all the way upstream in
enroll that removes more noisy issues like this by never publishing the
older plan years, while MA uses a more correct, historical
view.  As a result this only arises in MA currently, but we should
correct this everywhere.